@@ -25,7 +25,7 @@ from genshin.utility import auth as auth_utility
 
 __all__ = ["PAGES", "enter_code", "launch_webapp", "solve_geetest"]
 
-PAGES: typing.Final[typing.Dict[typing.Literal["captcha", "enter-code"], str]] = {
+PAGES: typing.Final[dict[typing.Literal["captcha", "enter-code"], str]] = {
     "captcha": """
     <!DOCTYPE html>
     <head>
@@ -159,7 +159,7 @@ async def launch_webapp(
 
     @routes.get("/mmt")
     async def mmt_endpoint(request: web.Request) -> web.Response:
-        return web.json_response(mmt.dict() if mmt else {})
+        return web.json_response(mmt.model_dump() if mmt else {})
 
     @routes.post("/send-data")
     async def send_data_endpoint(request: web.Request) -> web.Response:
