@@ -34,11 +34,7 @@ class GenshinException(Exception):
     original: str = ""
     msg: str = ""
 
-    def __init__(
-        self,
-        response: typing.Mapping[str, typing.Any] = {},
-        msg: typing.Optional[str] = None,
-    ) -> None:
+    def __init__(self, response: typing.Mapping[str, typing.Any] = {}, msg: typing.Optional[str] = None) -> None:
         self.retcode = response.get("retcode", self.retcode)
         self.original = response.get("message", "")
         self.msg = msg or self.msg or self.original
@@ -192,7 +188,7 @@ class WrongOTP(GenshinException):
 class GeetestError(GenshinException):
     """Geetest triggered during the battle chronicle API request."""
 
-    def __init__(self, response: typing.Dict[str, typing.Any]) -> None:
+    def __init__(self, response: dict[str, typing.Any]) -> None:
         super().__init__(response)
 
     msg = "Geetest triggered during the battle chronicle API request."
@@ -236,8 +232,8 @@ class VerificationCodeRateLimited(GenshinException):
     msg = "Too many verification code requests for the account."
 
 
-_TGE = typing.Type[GenshinException]
-_errors: typing.Dict[int, typing.Union[_TGE, str, typing.Tuple[_TGE, typing.Optional[str]]]] = {
+_TGE = type[GenshinException]
+_errors: dict[int, typing.Union[_TGE, str, tuple[_TGE, typing.Optional[str]]]] = {
     # misc hoyolab
     -100: InvalidCookies,
     -108: "Invalid language.",
@@ -249,10 +245,7 @@ _errors: typing.Dict[int, typing.Union[_TGE, str, typing.Tuple[_TGE, typing.Opti
     # database game record
     10101: TooManyRequests,
     10102: DataNotPublic,
-    10103: (
-        InvalidCookies,
-        "Cookies are valid but do not have a hoyolab account bound to them.",
-    ),
+    10103: (InvalidCookies, "Cookies are valid but do not have a hoyolab account bound to them."),
     10104: "Cannot view real-time notes of other users.",
     # calculator
     -500001: "Invalid fields in calculation.",
@@ -273,10 +266,7 @@ _errors: typing.Dict[int, typing.Union[_TGE, str, typing.Tuple[_TGE, typing.Opti
     -2016: RedemptionCooldown,
     -2017: RedemptionClaimed,
     -2018: RedemptionClaimed,
-    -2021: (
-        RedemptionException,
-        "Cannot claim codes for accounts with adventure rank lower than 10.",
-    ),
+    -2021: (RedemptionException, "Cannot claim codes for accounts with adventure rank lower than 10."),
     # rewards
     -5003: AlreadyClaimed,
     # chinese
@@ -296,13 +286,13 @@ _errors: typing.Dict[int, typing.Union[_TGE, str, typing.Tuple[_TGE, typing.Opti
     -202: IncorrectGamePassword,
 }
 
-ERRORS: typing.Dict[int, typing.Tuple[_TGE, typing.Optional[str]]] = {
-    retcode: ((exc, None) if isinstance(exc, type) else (GenshinException, exc) if isinstance(exc, str) else exc)
+ERRORS: dict[int, tuple[_TGE, typing.Optional[str]]] = {
+    retcode: (GenshinException, exc) if isinstance(exc, str) else exc if isinstance(exc, tuple) else (exc, None)
     for retcode, exc in _errors.items()
 }
 
 
-def raise_for_retcode(data: typing.Dict[str, typing.Any]) -> typing.NoReturn:
+def raise_for_retcode(data: dict[str, typing.Any]) -> typing.NoReturn:
     """Raise an equivalent error to a response.
 
     game record:
@@ -337,7 +327,7 @@ def raise_for_retcode(data: typing.Dict[str, typing.Any]) -> typing.NoReturn:
     raise GenshinException(data)
 
 
-def check_for_geetest(data: typing.Dict[str, typing.Any]) -> None:
+def check_for_geetest(data: dict[str, typing.Any]) -> None:
     """Check if geetest was triggered during the request and raise an error if so."""
     if data["retcode"] in GEETEST_RETCODES:
         raise GeetestError(data)
