@@ -2,13 +2,7 @@
 
 import typing
 
-if typing.TYPE_CHECKING:
-    import pydantic.v1 as pydantic
-else:
-    try:
-        import pydantic.v1 as pydantic
-    except ImportError:
-        import pydantic
+import pydantic
 
 __all__ = [
     "AppLoginResult",
@@ -31,8 +25,8 @@ class StokenResult(pydantic.BaseModel):
     mid: str
     token: str
 
-    @pydantic.root_validator(pre=True)
-    def _transform_result(cls, values: typing.Dict[str, typing.Any]) -> typing.Dict[str, typing.Any]:
+    @pydantic.model_validator(mode="before")
+    def _transform_result(cls, values: dict[str, typing.Any]) -> dict[str, typing.Any]:
         return {
             "aid": values["user_info"]["aid"],
             "mid": values["user_info"]["mid"],
@@ -45,11 +39,11 @@ class CookieLoginResult(pydantic.BaseModel):
 
     def to_str(self) -> str:
         """Convert the login cookies to a string."""
-        return "; ".join(f"{key}={value}" for key, value in self.dict().items())
+        return "; ".join(f"{key}={value}" for key, value in self.model_dump().items())
 
-    def to_dict(self) -> typing.Dict[str, str]:
+    def to_dict(self) -> dict[str, str]:
         """Convert the login cookies to a dictionary."""
-        return self.dict()
+        return self.model_dump()
 
 
 class QRLoginResult(CookieLoginResult):
@@ -126,8 +120,8 @@ class DeviceGrantResult(pydantic.BaseModel):
     game_token: str
     login_ticket: typing.Optional[str] = None
 
-    @pydantic.root_validator(pre=True)
-    def _str_to_none(cls, data: typing.Dict[str, typing.Union[str, None]]) -> typing.Dict[str, typing.Union[str, None]]:
+    @pydantic.model_validator(mode="before")
+    def _str_to_none(cls, data: dict[str, typing.Union[str, None]]) -> dict[str, typing.Union[str, None]]:
         """Convert empty strings to `None`."""
         for key in data:
             if data[key] == "" or data[key] == "None":
