@@ -2,18 +2,11 @@
 
 from __future__ import annotations
 
-import datetime
 import typing
 
-if typing.TYPE_CHECKING:
-    import pydantic.v1 as pydantic
-else:
-    try:
-        import pydantic.v1 as pydantic
-    except ImportError:
-        import pydantic
+import pydantic
 
-from genshin.models.model import Aliased, APIModel, Unique
+from genshin.models.model import Aliased, APIModel, DateTimeField
 
 __all__ = [
     "TeapotReplica",
@@ -52,14 +45,14 @@ class TeapotReplicaBlueprint(APIModel):
     is_invalid: bool
 
 
-class TeapotReplica(APIModel, Unique):
+class TeapotReplica(APIModel):
     """Genshin serenitea pot replica."""
 
     post_id: str
     title: str
     content: str
-    images: typing.List[str] = Aliased("imgs")
-    created_at: datetime.datetime
+    images: list[str] = Aliased("imgs")
+    created_at: DateTimeField
     stats: TeapotReplicaStats
     lang: str  # type: ignore
 
@@ -73,11 +66,11 @@ class TeapotReplica(APIModel, Unique):
     has_more_content: bool
     token: str
 
-    @pydantic.validator("images", pre=True)
+    @pydantic.field_validator("images", mode="before")
     def __extract_urls(cls, images: typing.Sequence[typing.Any]) -> typing.Sequence[str]:
         return [image if isinstance(image, str) else image["url"] for image in images]
 
-    @pydantic.validator("video", pre=True)
+    @pydantic.field_validator("video", mode="before")
     def __extract_url(cls, video: typing.Any) -> typing.Optional[str]:
         if isinstance(video, str):
             return video
