@@ -3,13 +3,7 @@
 import re
 import typing
 
-if typing.TYPE_CHECKING:
-    import pydantic.v1 as pydantic
-else:
-    try:
-        import pydantic.v1 as pydantic
-    except ImportError:
-        import pydantic
+import pydantic
 
 from genshin.models.honkai import battlesuit
 from genshin.models.model import Aliased, APIModel, Unique
@@ -51,8 +45,8 @@ class FullBattlesuit(battlesuit.Battlesuit):
     weapon: BattlesuitWeapon
     stigmata: typing.Sequence[Stigma] = Aliased("stigmatas")
 
-    @pydantic.root_validator(pre=True)
-    def __unnest_char_data(cls, values: typing.Dict[str, typing.Any]) -> typing.Dict[str, typing.Any]:
+    @pydantic.model_validator(mode="before")
+    def __unnest_char_data(cls, values: dict[str, typing.Any]) -> dict[str, typing.Any]:
         if isinstance(values.get("character"), typing.Mapping):
             values.update(values["character"])
 
@@ -60,10 +54,6 @@ class FullBattlesuit(battlesuit.Battlesuit):
 
         return values
 
-    @pydantic.validator("stigmata")
+    @pydantic.field_validator("stigmata")
     def __remove_unequipped_stigmata(cls, value: typing.Sequence[Stigma]) -> typing.Sequence[Stigma]:
         return [stigma for stigma in value if stigma.id != 0]
-
-
-# shuffle validators around because of nesting
-FullBattlesuit.__pre_root_validators__.reverse()
