@@ -89,7 +89,11 @@ async def cache():
 
     os.makedirs(".pytest_cache", exist_ok=True)
     with open(".pytest_cache/hoyo_cache.json", "w", encoding="utf-8") as file:
-        json.dump(cache, file, indent=4, ensure_ascii=False)
+        try:
+            json.dump(cache, file, indent=4, ensure_ascii=False)
+        except TypeError:
+            # Some objects are not serializable
+            pass
 
 
 @pytest.fixture(scope="session")
@@ -205,7 +209,7 @@ def pytest_addoption(parser: pytest.Parser):
     parser.addoption("--cooperative", action="store_true")
 
 
-def pytest_collection_modifyitems(items: typing.List[pytest.Item], config: pytest.Config):
+def pytest_collection_modifyitems(items: list[pytest.Item], config: pytest.Config):
     if config.option.cooperative:
         for item in items:
             if isinstance(item, pytest.Function) and asyncio.iscoroutinefunction(item.obj):
