@@ -1,3 +1,4 @@
+r"""
 import typing
 
 import pytest
@@ -8,10 +9,7 @@ import genshin
 class LiteralCharacter(genshin.models.BaseCharacter): ...
 
 
-LiteralCharacter.__pre_root_validators__ = LiteralCharacter.__pre_root_validators__[:-1]
-
 lang = "en-us"  # initiate local scope
-
 
 @pytest.mark.parametrize(
     ("data", "expected"),
@@ -119,7 +117,7 @@ genshin.models.APIModel.__new__ = APIModel___new__
 
 # def test_model_reserialization():
 #     for cls, model in sorted(all_models.items(), key=lambda pair: pair[0].__name__):
-#         cls(**model.dict())
+#         cls(**model.model_dump())
 
 #         if hasattr(model, "as_dict"):
 #             getattr(model, "as_dict")()
@@ -133,3 +131,4 @@ genshin.models.APIModel.__new__ = APIModel___new__
 #     os.makedirs(".pytest_cache", exist_ok=True)
 #     with open(".pytest_cache/hoyo_parsed.json", "w", encoding="utf-8") as file:
 #         file.write(data)
+"""
