@@ -1,15 +1,9 @@
 import enum
 import typing
 
-from genshin.models.model import Aliased, APIModel, Unique
+import pydantic
 
-if typing.TYPE_CHECKING:
-    import pydantic.v1 as pydantic
-else:
-    try:
-        import pydantic.v1 as pydantic
-    except ImportError:
-        import pydantic
+from genshin.models.model import Aliased, APIModel, Unique
 
 __all__ = (
     "AgentSkill",
@@ -137,7 +131,7 @@ class ZZZProperty(APIModel):
     type: typing.Union[int, ZZZPropertyType] = Aliased("property_id")
     value: str = Aliased("base")
 
-    @pydantic.validator("type", pre=True)
+    @pydantic.field_validator("type", mode="before")
     def __cast_id(cls, v: int) -> typing.Union[int, ZZZPropertyType]:
         # Prevent enum crash
         try:
